Optionally preserve the order of values
If TINYTOML_PRESERVE_TABLE_ORDER is defined a vector<pair<...>> is used
as the Table type instead of an unordered_map. This preserves the
insertion order.

The TOML spec is silent on the topic of element ordering but this can be
desirable for some files, e.g. when later tables should override the
content of earlier ones:

  [default]
  some_setting = 42

  [custom]
  some_setting = 17